api/requirements.txt
bcrypt==4.1.2
blinker==1.7.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.3
Flask==3.0.2
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector==2.2.9
paramiko==3.4.0
psycopg2-binary==2.9.9
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==1.0.1
redis==5.2.0
sshtunnel==0.4.0
Werkzeug==3.0.1
